rustam/requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.2
django-cleanup==6.0.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
fake-useragent==0.1.11
frozenlist==1.3.1
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
lxml==4.9.1
multidict==6.0.2
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.1
psycopg2-binary==2.9.4
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
selenium==4.5.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
yarl==1.8.1
zstandard==0.18.0
